Consolidate/combine computation queue item and computer merge objects into computations file.
